# openapi.yaml
openapi: 3.0.3
info:
  title: Building Server API
  description: RESTful API for user management with authentication and authorization
  version: 0.1.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://localhost:8443
    description: Production HTTPS server
  - url: http://localhost:8080
    description: Development HTTP server
...
components:
  schemas:
    Response:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
          description: Response message
        data:
          description: Optional response data
    ErrorResponse:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
          description: Error message
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Invalid or expired token
    Forbidden:
      description: Forbidden - Admin privileges required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Admin access required
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Validation error
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: User not found
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Email already exists
paths:
...
  /admin/books:
    post:
      tags:
        - Admin
      summary: Create new book
      description: Create a new book (admin only)
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: successfully created book
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /books/title/{title}:
    get:
      tags:
        - Book
      summary: Find books by title
      description: Retrieve books that match the given title
      operationId: getBooksByTitle
      parameters:
        - name: title
          in: path
          required: true
          description: Book title to search for
          schema:
            type: string
      responses:
        '200':
          description: Books retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookListResponse'
  /books/author/{author}:
    get:
      tags:
        - Book
      summary: Find books by author
      description: Retrieve books written by the given author
      operationId: getBooksByAuthor
      parameters:
        - name: author
          in: path
          required: true
          description: Author name to search for
          schema:
            type: string
      responses:
        '200':
          description: Books retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookListResponse'
  /books/id/{id}:
    get:
      tags:
        - Book
      summary: Get book by ID
      description: Retrieve detailed information for a specific book by ID
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          description: Book ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439012
      responses:
        '200':
          description: Book retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /books/borrow/{id}:
    post:
      tags:
        - Book
      summary: Borrow a book
      description: Borrow a book by ID (requires authentication)
      operationId: borrowBook
      parameters:
        - name: id
          in: path
          required: true
          description: Book ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439012
      responses:
        '200':
          description: Book borrowed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: successfully borrowed book
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /books/return/{id}:
    post:
      tags:
        - Book
      summary: Return a book
      description: Return a borrowed book by ID (requires authentication)
      operationId: returnBook
      parameters:
        - name: id
          in: path
          required: true
          description: Book ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439012
      responses:
        '200':
          description: Book returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: successfully returned book
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'